backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP client + scraping (no Playwright — httpx + BS4 only)
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Deduplication + data processing
rapidfuzz==3.10.0
pandas==2.2.3

# Resilience
tenacity==9.0.0

# Export
python-multipart==0.0.12
openpyxl==3.1.5

# LLM classification fallback
anthropic==0.40.0